# nip-backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API specification for the Authenticated Anonymity Architecture (AAA), a blockchain-based solution providing robust and ethical authenticated anonymous identities with deanonymization capabilities for criminal activity cases.",
        "title": "National Identity Provider (NIP)",
        "contact": {
            "name": "Michele Dinelli",
            "email": "m.dinelli@unibo.it"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8888",
    "basePath": "/",
    "paths": {
        "/v1/auth/pac": {
            "post": {
                "description": "Allows a user to request a Public Authentication Code (PAC). It accepts a payload `PID` and `SIGN(PID, SK)`, namely the PID signed with the private key of the public-private key pair used to obtain the PID. The PID must be sent as base64 (as it was retrieved originally from the NIP). The signed PID must be base64 encoded as well.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "PAC issuance request",
                "parameters": [
                    {
                        "description": "PAC Request Model",
                        "name": "models.PACRequestModel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PACRequestModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The Public Authentication Code",
                        "schema": {
                            "$ref": "#/definitions/models.PACResponseModel"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseModel"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseModel"
                        }
                    },
                    "500": {
                        "description": "An error occurred",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseModel"
                        }
                    }
                }
            }
        },
        "/v1/auth/sac": {
            "post": {
                "description": "The SAC (Secret Authentication Code) is a one-time code used to authenticate the user as an anonymous user. It accepts a payload `SID`, `ENC(SID, SK)`, namely the SID and the SID signed with the private key associated with the public key saved on the blockchain at the moment of seed phrase creation and used in the record where the SID was stored. The SID must be retrieved from the blockchain and sent as base64 encoded (supposedly from hex). The signed SID must be base64 encoded as well.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "SAC issuance request",
                "parameters": [
                    {
                        "description": "SAC Request Model",
                        "name": "models.SACRequestModel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SACRequestModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The Secret Authentication Code",
                        "schema": {
                            "$ref": "#/definitions/models.SACResponseModel"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseModel"
                        }
                    },
                    "500": {
                        "description": "An error occurred",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseModel"
                        }
                    }
                }
            }
        },
        "/v1/auth/verify-pac": {
            "post": {
                "description": "Allows a user to verify a PAC (Public Authentication Code) for services accepting an authenticated anonymous identity.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "PAC verification request",
                "parameters": [
                    {
                        "description": "PAC Verification Request Model",
                        "name": "models.PACVerificationRequestModel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PACVerificationRequestModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The Public Authentication Code Verification Response",
                        "schema": {
                            "$ref": "#/definitions/models.PACVerificationResponseModel"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseModel"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseModel"
                        }
                    },
                    "500": {
                        "description": "An error occurred",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseModel"
                        }
                    }
                }
            }
        },
        "/v1/identity/pid": {
            "post": {
                "description": "The NIP issues a Public Identity Data (PID), i.e., an anonymous token that identifies the user without sharing information. The real identity of the user is carried by the PID alone. The NIP will be the only one able to connect a PID to the real information of the respective user. It accepts a 2048 bit RSA PKCS#8 key encoded in base64.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "identity"
                ],
                "summary": "PID issuance request",
                "parameters": [
                    {
                        "description": "PID Request Model",
                        "name": "models.PIDRequestModel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PIDRequestModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The PID",
                        "schema": {
                            "$ref": "#/definitions/models.PIDResponseModel"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseModel"
                        }
                    },
                    "500": {
                        "description": "An error occurred",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ErrorResponseModel": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.PACRequestModel": {
            "type": "object",
            "required": [
                "pid",
                "signed_pid"
            ],
            "properties": {
                "pid": {
                    "type": "string",
                    "example": "DmIq8x2JNs+E2qGBr16LxP6lUK+i/5nJRLuzvYEymtY="
                },
                "signed_pid": {
                    "type": "string",
                    "example": "Tn7dpsKMqbhC3j8/o7OH5juVrkejhhZYkDHNqVyp8Am8uRktEDKO5i09iH9GZO0NOQHRRKD6lpwD0wC5GEPiuHN6aAof3WtPO9bqEB8ZLup4FjnyDh3mMX1UlvlYjKi6eTLMD2J8dwObz0nkmZOVsjk51o8jMapJtfnzMkYhADh4vZVDLpWYWbtBsmmXhNHS4SuWc+K3ZKMLtCYE/MFK3JT+Zcyabmrd9jmqvHRgi9XT+kABZ6XnyUSp4VWo1M7pl767/hZM5CxqgXORUMk8z9M9lHCDOsLfjBOwi81ObMtd+4oVgXgAbzBBHXnicB6X5PEBb5Qyh0RNyWFFNOEvtA=="
                }
            }
        },
        "models.PACResponseModel": {
            "type": "object",
            "properties": {
                "expiration": {
                    "type": "string",
                    "example": "2024-12-31T23:59:59Z"
                },
                "pac": {
                    "type": "integer",
                    "example": 874532
                }
            }
        },
        "models.PACVerificationRequestModel": {
            "type": "object",
            "required": [
                "pac",
                "pid"
            ],
            "properties": {
                "pac": {
                    "type": "integer",
                    "example": 874532
                },
                "pid": {
                    "type": "string",
                    "example": "DmIq8x2JNs+E2qGBr16LxP6lUK+i/5nJRLuzvYEymtY="
                }
            }
        },
        "models.PACVerificationResponseModel": {
            "type": "object",
            "properties": {
                "expiration": {
                    "type": "string",
                    "example": "2024-12-31T23:59:59Z"
                },
                "valid": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.PIDRequestModel": {
            "type": "object",
            "required": [
                "public_key"
            ],
            "properties": {
                "public_key": {
                    "type": "string",
                    "example": "LS0tLS1CRUdJTiBQVUJMSUMgS0VZLS0tLS0KTUlJQklqQU5CZ2txaGtpRzl3MEJBUUVGQUFPQ0FROEFNSUlCQ2dLQ0FRRUF1aVRTYzZMY3NmaFRZbEY0WDk2OApEcStXVzNrRmx3U21UR2NJQXJ5eFZ2OVhBQ2VFSXJuL0JYZE9McTFNSU52enc5WTZvNmhTd3hqUkp2SzlvdFpMCnlXYzFvc2RUNGdlTDJaT2xEMHUwakFvbEpqVU9hVWVyMmJKQTUzQzhXMjd1MHJyQzVvc3Z3THRNVWYxWnpTR0MKL2hrY0VOcjNjRGl6YWJibkw2dzVFUEdCVjlNU2RnaUdlelJXaXVEeGIzcldEOFNwSlpNa2c5b3ZSK3dOMlI2YwpOK2RwTkNjdDdWYjQxaWhXU0VXT1o2djhFdTREV3FjVW0vQzBpLzZ3eVJkWU93Y0E3NktkMmkxcnNZdHFlQS84CnZuMklTTEp2eGdkM3ZCdXY1ZnFkS0kzRm0rd3NIVDhHdHlkQUw5VjF4V3g4bnFlcFlwMlIvaXJMNHFUeGFnVzkKMXdJREFRQUIKLS0tLS1FTkQgUFVCTElDIEtFWS0tLS0tCg=="
                }
            }
        },
        "models.PIDResponseModel": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "PID successfully issued and saved"
                },
                "pid": {
                    "type": "string",
                    "example": "Nifl3y+2jmuAxF26jqpjogu0ZYnA6IxSikjmTnnjm7k="
                }
            }
        },
        "models.SACRequestModel": {
            "type": "object",
            "required": [
                "sid",
                "signed_sid"
            ],
            "properties": {
                "sid": {
                    "type": "string",
                    "example": "MHhmMzZmMmRlNWFmNmVkODg0ODM3NTgwMmZjOWIxNWMyNzBkN2NmNWUwNmI0MTYxZGY2ZGM5YjYyNWZhMmU4NzE2"
                },
                "signed_sid": {
                    "type": "string",
                    "example": "lOY7KkjuT3HV7R0/l66LR/0vR5URqy6i0cMH4nEHebks4ZSIHPO63FY+DCZX9QRu3VPt/eO21Fwm69pXk1lWbETll5vXbpChkKjaOAsuJfIRqWBX+bXHfiBCPvdjGr0f5YfWDx8xL1ndbXntNYF1WBt02Z/2JG06XsPsYwrT1FFDw6GiJiliJrBBJ7CYj4GhgdrgD70ZPdadKl7ChAgp9erc9s1nZ3Kdhnp9xmUR+aHcxi0RfikCvhgZvVvbHneSviI8tjp8zxT4VJhX2gbIJ/0fp4CVIaiHxHsMzCPB3rFdNNo5Pqr4TyDG/IP5k3QLD1yQXY+a9wlw6MGUBEiXrw=="
                }
            }
        },
        "models.SACResponseModel": {
            "type": "object",
            "properties": {
                "expiration": {
                    "type": "string",
                    "example": "2024-12-31T23:59:59Z"
                },
                "sac": {
                    "type": "integer",
                    "example": 874532
                }
            }
        }
    }
}